Refactor namespaces to improve clarity and align with package structure
