+++ src/Exception/PimcoreHttpCacheException.php
@@ -3,7 +3,7 @@
 namespace Neusta\Pimcore\HttpCacheBundle\Exception;
 
 /**
- *  Interface for all exceptions thrown by the PimcoreHttpCacheBundle.
+ * Interface for all exceptions thrown by the PimcoreHttpCacheBundle.
  */
 interface PimcoreHttpCacheException
 {
