+++ src/ResponseTagger/OnlyWhenActiveResponseTagger.php
@@ -1,10 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace Neusta\Pimcore\HttpCacheBundle\Cache\ResponseTagger;
+namespace Neusta\Pimcore\HttpCacheBundle\ResponseTagger;
 
-use Neusta\Pimcore\HttpCacheBundle\Cache\CacheTags;
-use Neusta\Pimcore\HttpCacheBundle\Cache\ResponseTagger;
 use Neusta\Pimcore\HttpCacheBundle\CacheActivator;
+use Neusta\Pimcore\HttpCacheBundle\CacheTags;
+use Neusta\Pimcore\HttpCacheBundle\ResponseTagger;
 
 final class OnlyWhenActiveResponseTagger implements ResponseTagger
 {
